Some struggling to get blinking R0Is to behave correctly. Still not at all perfect.
git-svn-id: file:///Users/wmohler/Documents/CytoSHOWsvnClone/svn_load_compat@19780 d59c7341-666b-7e48-8bb3-2014e8cfb900

+++ GLOWormJ022614/GLOWormJviaperlSVN/GLOWormJviaperlSVN32last/src/ij/gui/Roi.java
@@ -1454,9 +1454,9 @@ public class Roi extends Object implements Cloneable, java.io.Serializable {
 		if (width==0)
 			stroke = null;
 		else if (wideLine)
-			this.stroke = new BasicStroke(width, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL);
+			this.stroke = new BasicStroke((float)width, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL);
 		else
-			this.stroke = new BasicStroke(width);
+			this.stroke = new BasicStroke((float)width);
 		if (width>1f) fillColor = null;
 	}
